Fix MSVC build: rename LogLevel enum to PascalCase, silence narrowing warnings
* Rename LogLevel::{DEBUG,INFO,WARN,ERROR} -> {Debug,Info,Warn,Error} so the
  enum no longer collides with the ERROR macro that <wingdi.h> defines as 0,
  which broke compilation of tools/emission_tool.cpp on MSVC (the only
  translation unit that includes <windows.h> before logger.h).
* In src/main.cpp explicitly narrow nlohmann::json::value<int>() back to
  uint8_t when reading the time window, silencing /W4 C4244 warnings.

Linux + tests still green (41/41).

+++ src/main.cpp
@@ -49,10 +49,14 @@ AppConfig load_config(const fs::path& path) {
     cfg.poll_timeout_ms = doc.value("poll_timeout_ms", cfg.poll_timeout_ms);
     if (doc.contains("window")) {
         const auto& w = doc["window"];
-        cfg.window_start.hour = w.value("start_hour", static_cast<int>(cfg.window_start.hour));
-        cfg.window_start.minute = w.value("start_minute", static_cast<int>(cfg.window_start.minute));
-        cfg.window_end.hour = w.value("end_hour", static_cast<int>(cfg.window_end.hour));
-        cfg.window_end.minute = w.value("end_minute", static_cast<int>(cfg.window_end.minute));
+        cfg.window_start.hour = static_cast<uint8_t>(
+            w.value("start_hour", static_cast<int>(cfg.window_start.hour)));
+        cfg.window_start.minute = static_cast<uint8_t>(
+            w.value("start_minute", static_cast<int>(cfg.window_start.minute)));
+        cfg.window_end.hour = static_cast<uint8_t>(
+            w.value("end_hour", static_cast<int>(cfg.window_end.hour)));
+        cfg.window_end.minute = static_cast<uint8_t>(
+            w.value("end_minute", static_cast<int>(cfg.window_end.minute)));
     }
     return cfg;
 }
